Add files via upload
This Mathematica code conducts a weakly nonlinear analysis for the localised bulging of an inflated membrane tube of finite length. It is part of a longer code that serves two purposes. Firstly, it shows how accurate the 1D model of Lestringant & Audoly (Proc. R. Soc 2018) is. Secondly, it demonstrates that the tube does not need to be very long for the infinite length assumption to be valid.